contrib: gunyah-hypervisor update 12/2023
Gunyah updates:
* cyclomatic complexity improvements
* Misra fixes
* trbe rename
* virtio-input device handling
* remote cpu halt on debug
* memdb bitmap object ID count increase to 16
* provide vcpu_runtime_flags bitfield
* Return cause of VCPU sleep from vcpu_run
* vgic hw_active tracking fixes
* vcpu dying power-on race fix
* vcpu power voting fixes
* vgic unbind fixes
* idle with wfi timeout support
* wait for slow irqs after cpu wake
* page table barriers for correct walk cache flushing
* vgic 1:N routing fixes
* defer per-cpu gic initialization
* Avoid unsafe casts in atomic bitfield ops
* new platform functional cpus api

+++ hyp/core/partition_standard/src/init.c
@@ -45,9 +45,9 @@ static const paddr_t   phys_last  = (paddr_t)&image_phys_last;
 
 #if defined(ARCH_ARM) && ARCH_IS_64BIT
 // Ensure hypervisor is 2MiB page size aligned to use AArch64 2M block mappings
-static_assert((PLATFORM_RW_DATA_SIZE & 0x1fffffU) == 0U,
+static_assert(((size_t)PLATFORM_RW_DATA_SIZE & 0x1fffffU) == 0U,
 	      "PLATFORM_RW_DATA_SIZE must be 2MB aligned");
-static_assert((PLATFORM_HEAP_PRIVATE_SIZE & 0xfffU) == 0U,
+static_assert(((size_t)PLATFORM_HEAP_PRIVATE_SIZE & 0xfffU) == 0U,
 	      "PLATFORM_HEAP_PRIVATE_SIZE must be 4KB aligned");
 #endif
